tests: move URL-classification tests to test_handlers
_classify_url is defined in handlers.py, so testing it from
test_download.py mixed unrelated contexts. Move the two tests
to test_handlers.py alongside the other handler URL-pattern tests.
